refactor: split related topics into related software and projects sections on software and project edit pages tests: update e2e and unit tests
