+++ e2e/helpers/utils.ts
@@ -103,104 +103,6 @@ export async function openEditSection(page:Page,name:string) {
 
 }
 
-// TODO! REMOVE THIS FUNCTION LATER
-// export async function addOrganisation(page, organisation: Organisation, apiUrl) {
-
-//   // const findOrganisation = page.getByLabel('Find or add organisation')
-//   const findOrganisation = page.getByRole('combobox', {name: 'Find or add organisation'})
-
-//   // check if no organisation message is present
-//   const alert = await page.getByRole('alert')
-//     .filter({
-//       hasText: /No participating organisations/
-//     }).count() > 0
-
-//   // set breakpoint
-//   // console.log('alert...', alert)
-//   // await page.pause()
-
-//   if (alert === false) {
-//     //check if organisation already present
-//     const organisations = page.getByTestId('organisation-list-item')
-//     const [count] = await Promise.all([
-//       organisations.count(),
-//       page.waitForLoadState('networkidle')
-//     ])
-
-//     if (count > 0) {
-//       // check if organisation existis
-//       const found = await organisations
-//         .filter({hasText: RegExp(organisation.name)})
-//         .count()
-//       // console.log('found...', found)
-//       // if already exists we return false
-//       if (found > 0) return false
-//     }
-//   }
-
-//   // if not exists we search
-//   await Promise.all([
-//     page.waitForResponse(/api.ror.org\/organizations/),
-//     page.waitForLoadState('networkidle'),
-//     findOrganisation.fill(organisation.name)
-//   ])
-
-//   const options = page.getByTestId('find-organisation-option')
-//   const option = await options
-//     .filter({
-//       hasText:RegExp(organisation.name,'i')
-//     })
-//     .first()
-
-//   // get source information
-//   const source = await option.getByTestId('organisation-list-item').getAttribute('data-source')
-
-//   if (source === 'RSD') {
-//     await Promise.all([
-//       page.waitForResponse(RegExp(apiUrl)),
-//       option.click()
-//     ])
-//     // if rsd we just add it to list (no modal popup)
-//     return true
-//   }
-//   if (source === 'ROR') {
-//     // for ROR we can upload logo
-//     // in the modal/dialog so
-//     // we wait on modal to appear
-//     await Promise.all([
-//       page.waitForSelector('[role="dialog"]'),
-//       option.click()
-//     ])
-
-//     // console.log('logo...', organisation.logo)
-//     // upload logo if logo provided
-//     if (organisation.logo) {
-//       await uploadFile(
-//         page, '#upload-avatar-image',
-//         organisation.logo,
-//         'img'
-//       )
-//     }
-//     // save new organisation
-//     const saveBtn = page.getByRole('button', {
-//       name: 'Save'
-//     })
-//     await Promise.all([
-//       page.waitForResponse(/organisation/),
-//       page.waitForResponse(RegExp(apiUrl)),
-//       saveBtn.click()
-//     ])
-//   }
-
-//   // validate item is added to list
-//   const lastItem = await page.getByTestId('organisation-list-item').last()
-//   const lastText = await lastItem.getByTestId('organisation-list-item-text').textContent()
-//   // console.log('lastText...', lastText)
-//   expect(lastText).toContain(organisation.name)
-
-//   return true
-// }
-
 export async function addRelatedSoftware(page: Page, waitForResponse:string) {
   const findSoftware = page
     .getByTestId('find-related-software')
