+++ frontend/components/projects/edit/related-projects/FindRelatedProject.tsx
@@ -1,6 +1,8 @@
 // SPDX-FileCopyrightText: 2022 Dusan Mijatovic (dv4all)
 // SPDX-FileCopyrightText: 2022 Dusan Mijatovic (dv4all) (dv4all)
 // SPDX-FileCopyrightText: 2022 dv4all
+// SPDX-FileCopyrightText: 2023 Dusan Mijatovic (Netherlands eScience Center)
+// SPDX-FileCopyrightText: 2023 Netherlands eScience Center
 //
 // SPDX-License-Identifier: Apache-2.0
 
@@ -61,7 +63,7 @@ export default function FindRelatedProject({project,config,token,onAdd, onCreate
   }
 
   function onAddSelected(selected:AutocompleteOption<SearchProject>) {
-    if (selected && selected.data) {
+    if (selected?.data) {
       onAdd(selected.data)
     }
   }
@@ -131,7 +133,7 @@ export default function FindRelatedProject({project,config,token,onAdd, onCreate
           // if onCreate fn is not provided
           // we do not allow free solo text
           // eg. only selection of found items
-          freeSolo: onCreate ? true : false
+          freeSolo: typeof(onCreate) !== 'undefined'
         }}
       />
     </section>
