+++ frontend/components/projects/edit/related-projects/__mocks__/relatedProjectsForProject.json.license
@@ -1,0 +1,7 @@
+SPDX-FileCopyrightText: 2023 Dusan Mijatovic (Netherlands eScience Center)
+SPDX-FileCopyrightText: 2023 Dusan Mijatovic (dv4all)
+SPDX-FileCopyrightText: 2023 Dusan Mijatovic (dv4all) (dv4all)
+SPDX-FileCopyrightText: 2023 Netherlands eScience Center
+SPDX-FileCopyrightText: 2023 dv4all
+
+SPDX-License-Identifier: Apache-2.0
