+++ frontend/components/projects/edit/related-projects/index.tsx
@@ -8,18 +8,19 @@
 import {useEffect, useState} from 'react'
 
 import {useSession} from '~/auth'
-import {cfgRelatedItems as config} from './config'
 import {getRelatedProjectsForProject} from '~/utils/getProjects'
 import {addRelatedProject, deleteRelatedProject} from '~/utils/editProject'
-import useSnackbar from '~/components/snackbar/useSnackbar'
 import {sortOnStrProp} from '~/utils/sortFn'
+import {extractErrorMessages} from '~/utils/fetchHelpers'
 import {ProjectStatusKey, RelatedProjectForProject, SearchProject} from '~/types/Project'
+import {Status} from '~/types/Organisation'
+import useSnackbar from '~/components/snackbar/useSnackbar'
+import EditSectionTitle from '~/components/layout/EditSectionTitle'
+import EditSection from '~/components/layout/EditSection'
+import {relatedProject as config} from './config'
 import FindRelatedProject from './FindRelatedProject'
 import useProjectContext from '../useProjectContext'
 import RelatedProjectList from './RelatedProjectList'
-import EditSectionTitle from '~/components/layout/EditSectionTitle'
-import {Status} from '~/types/Organisation'
-import {extractErrorMessages} from '~/utils/fetchHelpers'
 
 export default function RelatedProjectsForProject() {
   const {token} = useSession()
@@ -124,35 +125,41 @@ export default function RelatedProjectsForProject() {
   }
 
   return (
-    <>
-      <EditSectionTitle
-        title={config.relatedProject.title}
-        subtitle={config.relatedProject.subtitle}
-      >
-        {/* add count to title */}
-        {relatedProject && relatedProject.length > 0 ?
-          <div className="pl-4 text-2xl">{relatedProject.length}</div>
-          : null
-        }
-      </EditSectionTitle>
-      <FindRelatedProject
-        project={project.id}
-        token={token}
-        config={{
-          freeSolo: false,
-          minLength: config.relatedProject.validation.minLength,
-          label: config.relatedProject.label,
-          help: config.relatedProject.help,
-          reset: true
-        }}
-        onAdd={onAdd}
-      />
-      <div className="py-8">
+    <EditSection className="flex-1 md:flex md:flex-col-reverse md:justify-end xl:grid xl:grid-cols-[3fr,2fr] xl:px-0 xl:gap-[3rem]">
+      <section className="py-4">
+        <EditSectionTitle
+          title={config.title}
+          // subtitle={config.subtitle}
+        >
+          {/* add count to title */}
+          {relatedProject && relatedProject.length > 0 ?
+            <div className="pl-4 text-2xl">{relatedProject.length}</div>
+            : null
+          }
+        </EditSectionTitle>
         <RelatedProjectList
           projects={relatedProject}
           onRemove={onRemove}
         />
-      </div>
-    </>
+      </section>
+      <section className="py-4">
+        <EditSectionTitle
+          title={config.findTitle}
+          // subtitle={config.label}
+        />
+        <FindRelatedProject
+          project={project.id}
+          token={token}
+          config={{
+            freeSolo: false,
+            minLength: config.validation.minLength,
+            label: config.label,
+            help: config.help,
+            reset: true
+          }}
+          onAdd={onAdd}
+        />
+      </section>
+    </EditSection>
   )
 }
