+++ frontend/components/projects/edit/related-software/config.ts
@@ -1,0 +1,18 @@
+// SPDX-FileCopyrightText: 2022 Dusan Mijatovic (dv4all)
+// SPDX-FileCopyrightText: 2022 dv4all
+// SPDX-FileCopyrightText: 2023 Dusan Mijatovic (Netherlands eScience Center)
+// SPDX-FileCopyrightText: 2023 Netherlands eScience Center
+//
+// SPDX-License-Identifier: Apache-2.0
+
+export const relatedSoftware = {
+  title: 'Related software',
+  subtitle: 'This project uses the following software registered in RSD',
+  findTitle: 'Find software',
+  label: 'Search for related software in RSD',
+  help: 'Start typing for suggestions',
+  validation: {
+    //custom validation rule, not in used by react-hook-form
+    minLength: 1,
+  }
+}
