+++ frontend/components/software/edit/related-projects/config.ts
@@ -1,0 +1,18 @@
+// SPDX-FileCopyrightText: 2022 Dusan Mijatovic (dv4all)
+// SPDX-FileCopyrightText: 2022 dv4all
+// SPDX-FileCopyrightText: 2023 Dusan Mijatovic (Netherlands eScience Center)
+// SPDX-FileCopyrightText: 2023 Netherlands eScience Center
+//
+// SPDX-License-Identifier: Apache-2.0
+
+export const relatedProject = {
+  title: 'Related projects',
+  subtitle: 'This project is related to these projects registered in RSD',
+  findTitle: 'Find project',
+  label: 'Find related projects in RSD',
+  help: 'Start typing for suggestions',
+  validation: {
+    //custom validation rule, not in used by react-hook-form
+    minLength: 1,
+  }
+}
