+++ frontend/components/software/edit/related-projects/index.tsx
@@ -1,22 +1,25 @@
 // SPDX-FileCopyrightText: 2022 - 2023 Dusan Mijatovic (dv4all)
 // SPDX-FileCopyrightText: 2022 - 2023 dv4all
+// SPDX-FileCopyrightText: 2023 Dusan Mijatovic (Netherlands eScience Center)
+// SPDX-FileCopyrightText: 2023 Netherlands eScience Center
 //
 // SPDX-License-Identifier: Apache-2.0
 
 import {useEffect, useState} from 'react'
 
 import {useSession} from '~/auth'
-import {cfgRelatedItems as config} from './config'
 import {getRelatedProjectsForSoftware} from '~/utils/getSoftware'
 import {addRelatedSoftware, deleteRelatedSoftware} from '~/utils/editProject'
-import useSnackbar from '~/components/snackbar/useSnackbar'
 import {sortOnStrProp} from '~/utils/sortFn'
-import {SearchProject} from '~/types/Project'
-import useSoftwareContext from '../useSoftwareContext'
-import FindRelatedProject from '~/components/projects/edit/related/FindRelatedProject'
-import EditSectionTitle from '~/components/layout/EditSectionTitle'
-import RelatedProjectList from '~/components/projects/edit/related/RelatedProjectList'
 import {Status} from '~/types/Organisation'
+import {SearchProject} from '~/types/Project'
+import useSnackbar from '~/components/snackbar/useSnackbar'
+import FindRelatedProject from '~/components/projects/edit/related-projects/FindRelatedProject'
+import EditSectionTitle from '~/components/layout/EditSectionTitle'
+import RelatedProjectList from '~/components/projects/edit/related-projects/RelatedProjectList'
+import EditSection from '~/components/layout/EditSection'
+import useSoftwareContext from '../useSoftwareContext'
+import {relatedProject as config} from './config'
 
 export default function RelatedProjectsForSoftware() {
   const {token} = useSession()
@@ -34,12 +37,11 @@ export default function RelatedProjectsForSoftware() {
         frontend: true,
         approved: false
       })
-      // extract software object
-      const projects = resp
-        .sort((a, b) => sortOnStrProp(a, b, 'title'))
+      // order on title
+      resp.sort((a, b) => sortOnStrProp(a, b, 'title'))
       if (abort) return null
       // debugger
-      setRelatedProject(projects)
+      setRelatedProject(resp)
       setLoadedSoftware(software?.id ?? '')
     }
     if (software.id && token &&
@@ -102,35 +104,41 @@ export default function RelatedProjectsForSoftware() {
   }
 
   return (
-    <>
-      <EditSectionTitle
-        title={config.relatedProject.title}
-        subtitle={config.relatedProject.subtitle}
-      >
-        {/* add count to title */}
-        {relatedProject && relatedProject.length > 0 ?
-          <div className="pl-4 text-2xl">{relatedProject.length}</div>
-          : null
-        }
-      </EditSectionTitle>
-      <FindRelatedProject
-        project={''}
-        token={token}
-        config={{
-          freeSolo: false,
-          minLength: config.relatedProject.validation.minLength,
-          label: config.relatedProject.label,
-          help: config.relatedProject.help,
-          reset: true
-        }}
-        onAdd={onAdd}
-      />
-      <div className="py-8">
+    <EditSection className="flex-1 md:flex md:flex-col-reverse md:justify-end xl:grid xl:grid-cols-[3fr,2fr] xl:px-0 xl:gap-[3rem]">
+      <section className="py-4">
+        <EditSectionTitle
+          title={config.title}
+          // subtitle={config.subtitle}
+        >
+          {/* add count to title */}
+          {relatedProject && relatedProject.length > 0 ?
+            <div className="pl-4 text-2xl">{relatedProject.length}</div>
+            : null
+          }
+        </EditSectionTitle>
         <RelatedProjectList
           projects={relatedProject}
           onRemove={onRemove}
         />
-      </div>
-    </>
+      </section>
+      <section className="py-4">
+        <EditSectionTitle
+          title={config.findTitle}
+          // subtitle={config.label}
+        />
+        <FindRelatedProject
+          project={''}
+          token={token}
+          config={{
+            freeSolo: false,
+            minLength: config.validation.minLength,
+            label: config.label,
+            help: config.help,
+            reset: true
+          }}
+          onAdd={onAdd}
+        />
+      </section>
+    </EditSection>
   )
 }
