+++ frontend/components/software/edit/related-software/__mocks__/relatedSoftwareForSoftware.json.license
@@ -1,4 +1,6 @@
+SPDX-FileCopyrightText: 2023 Dusan Mijatovic (Netherlands eScience Center)
 SPDX-FileCopyrightText: 2023 Dusan Mijatovic (dv4all) (dv4all)
+SPDX-FileCopyrightText: 2023 Netherlands eScience Center
 SPDX-FileCopyrightText: 2023 dv4all
 
 SPDX-License-Identifier: Apache-2.0
