+++ frontend/components/software/edit/related-software/index.tsx
@@ -1,23 +1,24 @@
 // SPDX-FileCopyrightText: 2022 Dusan Mijatovic (dv4all)
 // SPDX-FileCopyrightText: 2022 dv4all
+// SPDX-FileCopyrightText: 2023 Dusan Mijatovic (Netherlands eScience Center)
+// SPDX-FileCopyrightText: 2023 Netherlands eScience Center
 //
 // SPDX-License-Identifier: Apache-2.0
 
 import {useEffect, useState} from 'react'
 
 import {useSession} from '~/auth'
-import {RelatedSoftwareOfSoftware, SearchSoftware} from '~/types/SoftwareTypes'
-
-import {cfgRelatedItems as config} from './config'
-import useSnackbar from '~/components/snackbar/useSnackbar'
 import {sortOnStrProp} from '~/utils/sortFn'
-import useSoftwareContext from '../useSoftwareContext'
-import FindRelatedSoftware from '~/components/projects/edit/related/FindRelatedSoftware'
 import {addRelatedSoftware, deleteRelatedSoftware, getRelatedSoftwareForSoftware} from '~/utils/editRelatedSoftware'
-
-import RelatedSoftwareList from '../../../projects/edit/related/RelatedSoftwareList'
-import EditSectionTitle from '~/components/layout/EditSectionTitle'
+import {RelatedSoftwareOfSoftware, SearchSoftware} from '~/types/SoftwareTypes'
 import {Status} from '~/types/Organisation'
+import useSnackbar from '~/components/snackbar/useSnackbar'
+import FindRelatedSoftware from '~/components/projects/edit/related-software/FindRelatedSoftware'
+import RelatedSoftwareList from '~/components/projects/edit/related-software/RelatedSoftwareList'
+import EditSectionTitle from '~/components/layout/EditSectionTitle'
+import EditSection from '~/components/layout/EditSection'
+import useSoftwareContext from '../useSoftwareContext'
+import {relatedSoftware as config} from './config'
 
 
 export default function RelatedSoftwareForSoftware() {
@@ -109,35 +110,41 @@ export default function RelatedSoftwareForSoftware() {
   }
 
   return (
-    <>
-      <EditSectionTitle
-        title={config.relatedSoftware.title}
-        subtitle={config.relatedSoftware.subtitle}
-      >
-        {/* add count to title */}
-        {relatedSoftware && relatedSoftware.length > 0 ?
-          <div className="pl-4 text-2xl">{relatedSoftware.length}</div>
-          : null
-        }
-      </EditSectionTitle>
-      <FindRelatedSoftware
-        software={software.id ?? ''}
-        token={token}
-        config={{
-          freeSolo: false,
-          minLength: config.relatedSoftware.validation.minLength,
-          label: config.relatedSoftware.label,
-          help: config.relatedSoftware.help,
-          reset: true
-        }}
-        onAdd={onAdd}
-      />
-      <div className="py-8">
+    <EditSection className="flex-1 md:flex md:flex-col-reverse md:justify-end xl:grid xl:grid-cols-[3fr,2fr] xl:px-0 xl:gap-[3rem]">
+      <section className="py-4">
+        <EditSectionTitle
+          title={config.title}
+        // subtitle={config.subtitle}
+        >
+          {/* add count to title */}
+          {relatedSoftware && relatedSoftware.length > 0 ?
+            <div className="pl-4 text-2xl">{relatedSoftware.length}</div>
+            : null
+          }
+        </EditSectionTitle>
         <RelatedSoftwareList
           software={relatedSoftware}
           onRemove={onRemove}
         />
-      </div>
-    </>
+      </section>
+      <section className="py-4">
+        <EditSectionTitle
+          title={config.findTitle}
+        // subtitle={config.label}
+        />
+        <FindRelatedSoftware
+          software={software.id ?? ''}
+          token={token}
+          config={{
+            freeSolo: false,
+            minLength: config.validation.minLength,
+            label: config.label,
+            help: config.help,
+            reset: true
+          }}
+          onAdd={onAdd}
+        />
+      </section>
+    </EditSection>
   )
 }
