--- frontend/components/software/edit/related/__mocks__/relatedSoftwareForSoftware.json.license
@@ -1,4 +1,0 @@
-SPDX-FileCopyrightText: 2023 Dusan Mijatovic (dv4all) (dv4all)
-SPDX-FileCopyrightText: 2023 dv4all
-
-SPDX-License-Identifier: Apache-2.0
